+++ softgroup/services/authorization/authorization-impl/src/main/java/com/softgroup/authorization/impl/handler/LoginAuthorizationHandler.java
@@ -1,23 +1,23 @@
-package com.softgroup.authorithation.impl.handler;
+package com.softgroup.authorization.impl.handler;
 
 import com.softgroup.authorization.api.router.AuthorizationRequestHandler;
 import com.softgroup.common.protocol.Request;
 import com.softgroup.common.protocol.Response;
+import org.springframework.stereotype.Component;
 
 /**
  * Author: vadym
  * Date: 23.02.17
- * Time: 13:21
- */
-public class RegisterAuthorizationHandler implements AuthorizationRequestHandler {
+ * Time: 13:16
+ */@Component
+public class LoginAuthorizationHandler implements AuthorizationRequestHandler {
     @Override
     public String getName() {
-        return "registerHandler";
+        return "loginHandler";
     }
 
     @Override
     public Response<?> handle(Request<?> msg) {
-
         return null;
     }
 }
