+++ softgroup/services/authorization/authorization-impl/src/main/java/com/softgroup/authorization/impl/handler/RegisterAuthorizationHandler.java
@@ -1,22 +1,25 @@
-package com.softgroup.authorithation.impl.handler;
+package com.softgroup.authorization.impl.handler;
 
 import com.softgroup.authorization.api.router.AuthorizationRequestHandler;
 import com.softgroup.common.protocol.Request;
 import com.softgroup.common.protocol.Response;
+import org.springframework.stereotype.Component;
 
 /**
  * Author: vadym
  * Date: 23.02.17
- * Time: 12:43
+ * Time: 13:21
  */
-public class SMSAuthorizationHandler implements AuthorizationRequestHandler {
+@Component
+public class RegisterAuthorizationHandler implements AuthorizationRequestHandler {
     @Override
     public String getName() {
-        return "smsHandler";
+        return "registerHandler";
     }
 
     @Override
     public Response<?> handle(Request<?> msg) {
+
         return null;
     }
 }
